+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 ### 11.52
  - Add WPA3 attacks menu as hookable function to facilitate WPA3 existing plugins integration
+ - Improved adapter 802.11be standard support detection
+ - Improved Evil Twin fake APs to add 802.11be if supported by the adapter
 
 ### 11.51
  - ESSID stripping using invisible char to bypass Apple SSID grouping in Evil Twin attacks